+++ thriftrw/idl/parser.py
@@ -26,7 +26,7 @@ from ply import yacc
 
 from . import ast
 from .lexer import Lexer
-from .exceptions import ThriftParserError
+from ..errors import ThriftParserError
 
 
 __all__ = ['Parser']
@@ -443,7 +443,7 @@ class Parser(ParserSpec):
 
         :param input:
             String containing the text to be parsed.
-        :raises ThriftParserError:
+        :raises thriftrw.errors.ThriftParserError:
             For parsing errors.
         """
         return self._parser.parse(input, lexer=self._lexer, **kwargs)